Setup and configure annotate models gem (#2)
* add annotate gem into project

* execute `rails g annotate:install`

* execute annotate and update Ticket model

* Add GitHub workflow for annotating code

This commit adds a new GitHub workflow file named `annotate.yml`. The workflow is triggered on pushes and pull requests to the `main` branch. The workflow contains a job called `check-source-code-with-rubocop` which runs on the latest version of Ubuntu. The job installs Ruby and gems using the `ruby/setup-ruby` action with bundler caching enabled. It then checks out the code using the `actions/checkout` action. Finally, it runs the `annotate` command with the `--frozen` flag to ensure everything is properly annotated.