Add rspec and unit test for ticket model (#10)
* Add .byebug_history to .gitignore

* add rspec-rails gem into project

* execute `rails g rspec:install`

* config: add eager_load config to production and test environments

This commit adds the `eager_load` configuration option to the production and test environments of the application. In the `production.rb` file, `config.eager_load` is set to `true`, ensuring that classes and modules are preloaded. In the `test.rb` file, `config.eager_load` is set to `false`, indicating that preloading is not necessary for testing. This change allows for better performance and consistency in the different environments of the application.

* Update database.yml to include test environment configuration

* Create validation test for Ticket model email attribute